refactor: replace SearchBox to another directory
