+++ src/components/main-page/search-box/SearchBox.module.scss
